+++ README.md
@@ -25,9 +25,13 @@ Nextをざっくりいうと、
 - Reactベースのフレームワーク
 - /pages以下のフォルダ/ファイル構成がそのままWebページの構造になる
 
-です。基本/pages以下以外は触らなくて良い(自動生成されたlintや依存関係のファイル)です。
+です。
 
-開発時は上記のコマンドで動作を確認してください。本番環境にデプロイするには、単にGitHubのmainブランチに変更を適用すれば、amplifyがいい感じにやってくれます。
+基本的にはReactの開発手法に近いです。/src/components以下に汎用的な部品を置き、/src/stylesにCSSを置き、/pagesでインポートして使います。
+
+開発時は上記のコマンドでサーバーを起動して動作を確認してください。本番環境にデプロイするには、単にGitHubのmainブランチに変更を適用すれば、amplifyがうまいことやってくれます。
+
+もし変更したい場合、ブラウザのAWSコンソール上のAmplifyのメニューから設定してください。
 
 ## Learn More
