+++ src/components/Logout.tsx
@@ -1,4 +1,13 @@
+import { Button, useAuthenticator } from "@aws-amplify/ui-react";
 import { Auth } from "aws-amplify";
+
 export default function Logout() {
-    return <li><a onClick={() => Auth.signOut()} >ログアウト</a></li>
+  //Authenticator.Providerでラップされたコンポーネント(layoutで囲っているのでどこでも使える)でしか使えない
+  //認証周りを触るためのおまじない
+  useAuthenticator((context) => [context.route]); 
+  return (
+    <Button>
+      <a onClick={() => Auth.signOut()}>ログアウト</a>
+    </Button>
+  );
 }
